refactor!: split single_letter_names into one rule per file
The previous `single_letter_names` module bundled four independent
lints sharing a `Config` struct whose four fields each fed exactly
one lint. Splitting:

- Each lint now lives in its own `src/rules/<rule>.rs` with its own
  `Config` keyed by `perfectionist::<rule>`. The shared
  `is_single_ascii_letter` and `binding_ident` helpers move to
  `src/rules/common.rs`.
- Rule-specific helpers (`enclosing_short_trait_impl`,
  `is_trivial_wrapper`, the closure-call introspection) follow
  their sole consumer into the respective file.
- `CLAUDE.md` gains a "One rule per file, one Config per rule"
  section codifying the convention, including a pre-implementation
  check to split big rules whose sub-checks can be cleanly
  separated.

Breaking: dylint.toml users who set
`[perfectionist.single_letter_names]` must move each field under
its rule's own table.
