+++ src/rules.rs
@@ -1,3 +1,5 @@
+pub(crate) mod common;
+
 pub mod arc_rc_clone;
 pub mod derive_ordering;
 pub mod flat_module_pattern;
@@ -5,7 +7,10 @@ pub mod macro_argument_binding;
 pub mod macro_trailing_comma;
 pub mod non_exhaustive_error;
 pub mod prefer_raw_string;
-pub mod single_letter_names;
+pub mod single_letter_closure_param;
+pub mod single_letter_function_param;
+pub mod single_letter_generic;
+pub mod single_letter_let_binding;
 pub mod unicode_ellipsis_in_comments;
 pub mod unicode_ellipsis_in_panic_messages;
 pub mod unknown_perfectionist_lints;
